+++ examples/server-side-rendering/src/app/app.component.ts
@@ -4,28 +4,10 @@ import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { TransferState, makeStateKey } from '@angular/platform-browser';
 import { history } from 'instantsearch.js/es/lib/routers';
 import { simple } from 'instantsearch.js/es/lib/stateMappings';
-import { createSSRSearchClient } from 'angular-instantsearch';
-import * as qs from 'qs';
-import { Routing } from 'instantsearch.js/es/lib/InstantSearch';
-
-// TODO: put this inside the library, maybe inside InstantSearch.js?
-function parseServerRequest(
-  req: { url: string } | undefined,
-  { router, stateMapping }: Routing
-) {
-  if (!req) {
-    return undefined;
-  }
-
-  // this is for faking the browser API used by the router.
-  const location = { search: `?${req.url.split('?')[1]}` };
-  // TODO: this is "private" in IS.js, but since you can change the parseURL
-  // in historyRouter, we need to use it directly
-  // we can't call .read here, since that uses browser globals, unless it can be injected
-  const routeState = router.parseURL({ qsModule: qs, location });
-
-  return stateMapping.routeToState(routeState);
-}
+import {
+  createSSRSearchClient,
+  parseServerRequest,
+} from 'angular-instantsearch';
 
 @Component({
   selector: 'app-root',
